requirements.txt
tk
Pillow
pyttsx3
SpeechRecognition
wikipedia
pygame
